move exportSpans to common
